eima_ai/requirements.txt
# Jarvis-Granite Live Telemetry - Dependencies
# Version 2.1 | December 2025

# =============================================================================
# Web Framework
# =============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# =============================================================================
# Data Validation
# =============================================================================
pydantic>=2.5.0
pydantic-settings>=2.1.0

# =============================================================================
# HTTP Client
# =============================================================================
httpx>=0.25.0

# =============================================================================
# IBM Services
# =============================================================================
ibm-watson>=8.0.0
ibm-cloud-sdk-core>=3.18.0

# =============================================================================
# LangChain (LLM management)
# =============================================================================
langchain>=0.1.0
langchain-ibm>=0.1.0

# =============================================================================
# LiveKit (WebRTC transport)
# =============================================================================
livekit>=0.9.0
livekit-api>=0.4.0

# =============================================================================
# Retry Logic
# =============================================================================
tenacity>=8.2.0

# =============================================================================
# Configuration
# =============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0.0
